fix(runtime): add MethodUpdate dispatch, fix interface assertion, extract clearOutdatedCmd
- Add case domain.MethodUpdate to BeginExecution switch, dispatching BeginUpdate command (fixes state violation when updating outdated arrows)
- Remove interface assertion var _ = beginExecCmdWithNilExec{} from reactions_test.go (follows project convention)
- Extract clearOutdatedCmd type and methods to new file clear_outdated.go
